Transition from setup.py to setup.cfg
