+++ .github/workflows/stage.yml
@@ -65,7 +65,7 @@ jobs:
 
       - name: Install playbook dependencies
         run: |
-          python3 -m pip install docker
+          python3 -m pip install docker setuptools_scm
 
       - name: Build and stage AWX
         working-directory: awx
