+++ requirements/requirements_dev.txt
@@ -1,3 +1,4 @@
+build
 django-debug-toolbar==3.2.4
 django-rest-swagger
 # pprofile - re-add once https://github.com/vpelletier/pprofile/issues/41 is addressed
