+++ setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = awx
+author = Red Hat
+author_email = info@ansible.com
+version = attr: awx.get_version
+
+[options]
+packages =
+    awx
+zip_safe = False
+include_package_data = True
+
+[options.entry_points]
+console_scripts =
+    awx-manage = awx:manage
+awx.credential_plugins =
+    conjur = awx.main.credential_plugins.conjur:conjur_plugin
+    hashivault_kv = awx.main.credential_plugins.hashivault:hashivault_kv_plugin
+    hashivault_ssh = awx.main.credential_plugins.hashivault:hashivault_ssh_plugin
+    azure_kv = awx.main.credential_plugins.azure_kv:azure_keyvault_plugin
+    aim = awx.main.credential_plugins.aim:aim_plugin
+    centrify_vault_kv = awx.main.credential_plugins.centrify_vault:centrify_plugin
+    thycotic_dsv = awx.main.credential_plugins.dsv:dsv_plugin
+    thycotic_tss = awx.main.credential_plugins.tss:tss_plugin
