+++ tools/ansible/build.yml
@@ -5,7 +5,7 @@
   tasks:
     - name: Get version from SCM if not explicitly provided
       shell: |
-        python setup.py --version | cut -d + -f -1
+        python3 -m setuptools_scm | cut -d + -f -1
       args:
         chdir: '../../'
       register: setup_py_version
