+++ tools/scripts/scm_version.py
@@ -1,0 +1,4 @@
+from setuptools_scm import get_version
+
+version = get_version(root='../..', relative_to=__file__)
+print(version)
