+++ tox.ini
@@ -1,12 +1,17 @@
+[tox]
+isolated_build = True
+
 [testenv:linters]
 deps =
   make
   black
   flake8
+  setuptools-scm
   yamllint
 allowlist_externals = make
 setenv =
   BLACK_ARGS = --check
+  PYTHON = python3
 commands =
   make black
   flake8 awx awxkit awx_collection
